Add possibility to use field alias in relationship
When creating a single relationship to an entity, the toHandle is the name of the entity.
When creating multiple relationships to the same entity, multiple toHandles should be created with different names.
By using the 'as' field in de field config of many-to-one and one-to-one relationships this is made possible.